move actions to same test dir as checkpoints
