+++ tests/checkpoint/test_core_actions.py
